break v2_step.py into account_, application_, and other_
